some small improvements
