+++ lib/dai-plugin-mcd/src/CdpManager.js
@@ -10,7 +10,8 @@ import { castAsCurrency, stringToBytes, bytesToString } from './utils';
 import padStart from 'lodash/padStart';
 import { MDAI, ETH } from './index';
 const { CDP_MANAGER, CDP_TYPE, SYSTEM_DATA, QUERY_API } = ServiceRoles;
-import { parseFrobEvents } from './ManagedCdp';
+import BigNumber from 'bignumber.js';
+import { RAY } from './constants';
 
 export default class CdpManager extends LocalService {
   constructor(name = CDP_MANAGER) {
@@ -61,7 +62,7 @@ export default class CdpManager extends LocalService {
     const cdpIds = await this.getCdpIds(proxyAddress);
     const ilksAndUrns = await Promise.all(
       cdpIds.map(async c => {
-        const urn = await this.getUrn(c.id); //will add caching to this function
+        const urn = await this.getUrn(c.id);
         const ilk = stringToBytes(c.ilk);
         return { urn, ilk };
       })
@@ -69,7 +70,7 @@ export default class CdpManager extends LocalService {
     const events = await this.get(QUERY_API).getCdpEventsForArrayOfIlksAndUrns(
       ilksAndUrns
     );
-    return parseFrobEvents(events, this.get(CDP_TYPE));
+    return this.parseFrobEvents(events, this.get(CDP_TYPE));
   }
 
   @tracksTransactions
@@ -144,6 +145,42 @@ export default class CdpManager extends LocalService {
       this._getUrnPromises[id] = this._manager.urns(id);
     }
     return this._getUrnPromises[id];
+  }
+
+  parseFrobEvents(events) {
+    return events.map(e => {
+      const ilk = bytesToString(e.ilkName);
+      const currency = this.get(CDP_TYPE).getCdpType(null, ilk).currency;
+      const transactionHash = e.tx.transactionHash;
+      const rate = new BigNumber(e.ilk.nodes[0].rate.toString()).dividedBy(RAY);
+      const changeInCollateral = currency.wei(Math.abs(e.dink));
+      let collateralAction;
+      if (parseInt(e.dink) !== 0) {
+        collateralAction = parseInt(e.dink) > 0 ? 'lock' : 'free';
+      }
+      const dart = MDAI.wei(Math.abs(e.dart));
+      const changeInDebt = dart.times(rate);
+      let daiAction;
+      if (parseInt(e.dart) !== 0) {
+        daiAction = parseInt(e.dart) > 0 ? 'draw' : 'wipe';
+      }
+      const time = new Date(e.tx.era.iso);
+      const senderAddress = e.tx.txFrom;
+      const resultingCollateral = currency.wei(e.urn.nodes[0].ink);
+      const resultingDebt = MDAI.wei(e.urn.nodes[0].art);
+      return {
+        transactionHash,
+        changeInCollateral,
+        collateralAction,
+        changeInDebt,
+        daiAction,
+        ilk,
+        time,
+        senderAddress,
+        resultingCollateral,
+        resultingDebt
+      };
+    });
   }
 
   getIdBytes(id, prefix = true) {
