+++ lib/dai-plugin-mcd/src/CdpType.js
@@ -100,6 +100,7 @@ export default class CdpType {
     return Promise.all(
       prices.map(async e => {
         const price = this.currency.wei(e.val);
+        //todo: update this query to read the datetime directly from vdb once vdb is updated with that functionality
         const timestamp = (await this._web3Service.getBlock(e.blockNumber))
           .timestamp;
         const time = new Date(1000 * timestamp);
