+++ lib/dai-plugin-mcd/src/ManagedCdp.js
@@ -1,48 +1,9 @@
-import { castAsCurrency, stringToBytes, bytesToString } from './utils';
+import { castAsCurrency, stringToBytes } from './utils';
 import { tracksTransactionsWithOptions } from './utils/tracksTransactions';
 import { ServiceRoles } from './constants';
 import ethAbi from 'web3-eth-abi';
 import assert from 'assert';
 import { MDAI } from './index';
-import BigNumber from 'bignumber.js';
-import { RAY } from './constants';
-
-export function parseFrobEvents(events, cdpTypeService) {
-  return events.map(e => {
-    const ilk = bytesToString(e.ilkName);
-    const currency = cdpTypeService.getCdpType(null, ilk).currency;
-    const transactionHash = e.tx.transactionHash;
-    const rate = new BigNumber(e.ilk.nodes[0].rate.toString()).dividedBy(RAY);
-    const changeInCollateral = currency.wei(Math.abs(e.dink));
-    let collateralAction;
-    if (parseInt(e.dink) !== 0) {
-      //could use .isEqual on the currency object instead?
-      collateralAction = parseInt(e.dink) > 0 ? 'lock' : 'free';
-    }
-    const dart = MDAI.wei(Math.abs(e.dart));
-    const changeInDebt = dart.times(rate);
-    let daiAction;
-    if (parseInt(e.dart) !== 0) {
-      daiAction = parseInt(e.dart) > 0 ? 'draw' : 'wipe';
-    }
-    const time = new Date(e.tx.era.iso);
-    const senderAddress = e.tx.txFrom;
-    const resultingCollateral = currency.wei(e.urn.nodes[0].ink);
-    const resultingDebt = MDAI.wei(e.urn.nodes[0].art);
-    return {
-      transactionHash,
-      changeInCollateral,
-      collateralAction,
-      changeInDebt,
-      daiAction,
-      ilk,
-      time,
-      senderAddress,
-      resultingCollateral,
-      resultingDebt
-    };
-  });
-}
 
 export default class ManagedCdp {
   constructor(id, ilk, cdpManager) {
@@ -140,7 +101,10 @@ export default class ManagedCdp {
     const events = await this._cdpManager
       .get(ServiceRoles.QUERY_API)
       .getCdpEventsForIlkAndUrn(stringToBytes(this.ilk), urn);
-    return parseFrobEvents(events, this._cdpManager.get(ServiceRoles.CDP_TYPE));
+    return this._cdpManager.parseFrobEvents(
+      events,
+      this._cdpManager.get(ServiceRoles.CDP_TYPE)
+    );
   }
 
   //todo: add caching?
