+++ lib/dai-plugin-mcd/src/QueryApiMcdService.js
@@ -54,7 +54,6 @@ export default class QueryApi extends PublicService {
       }`;
   }
 
-  //TODO: add error handling
   async getCdpEventsForIlkAndUrn(ilkName, urn) {
     const query = '{' + this._buildFrobsQuery(ilkName, urn) + '}';
     const response = await getQueryResponse(this.serverUrl, query);
@@ -70,7 +69,7 @@ export default class QueryApi extends PublicService {
     query += '}';
     const response = await getQueryResponse(this.serverUrl, query);
     let events = [];
-    cdps.forEach((cdp, index) => {
+    cdps.forEach((_, index) => {
       events.push(response[`frobs${index}`].nodes);
     });
     const arr = [].concat.apply([], events); //flatten array
